+++ NEWS.md
@@ -1,6 +1,8 @@
 # farver (development version)
 
 * Added support for OKLab and OKLch (polar version of OKLab) colour spaces
+* Added `encode_native()` and `decode_native()` to convert back and forth 
+  between the internal R colour format.
 
 # farver 2.0.3